Templated the yml file in a way it's usable and can be referred elsewhere

+++ scripts/cleanup-managedimages.ps1
@@ -6,6 +6,7 @@ param(
     [String] [Parameter (Mandatory=$true)] $ResourceGroup,
     [String] [Parameter (Mandatory=$true)] $AgentsResourceGroup,
     [String] [Parameter (Mandatory=$true)] $VmssNameWindows,
+    [String] [Parameter (Mandatory=$true)] $VmssNameWindows2022,
     [String] [Parameter (Mandatory=$true)] $VmssNameUbuntu
 )
 
@@ -15,10 +16,11 @@ az account set -s $SubscriptionId
 $managedImages = az image list --resource-group $ResourceGroup --subscription $SubscriptionId --query "[].id" | Out-String | ConvertFrom-Json
 
 $windowsManagedImage = az vmss show --name $VmssNameWindows --resource-group $AgentsResourceGroup --query 'virtualMachineProfile.storageProfile.imageReference.id' --subscription $SubscriptionId | Out-String | ConvertFrom-Json
+$windowsManagedImage2022 = az vmss show --name $VmssNameWindows2022 --resource-group $AgentsResourceGroup --query 'virtualMachineProfile.storageProfile.imageReference.id' --subscription $SubscriptionId | Out-String | ConvertFrom-Json
 $ubuntuManagedImage = az vmss show --name $VmssNameUbuntu --resource-group $AgentsResourceGroup --query 'virtualMachineProfile.storageProfile.imageReference.id' --subscription $SubscriptionId | Out-String | ConvertFrom-Json
 
 foreach ($managedImage in $managedImages) {
-    if ($managedImage -ne $windowsManagedImage -and $managedImage -ne $ubuntuManagedImage) {
+    if ($managedImage -ne $windowsManagedImage -and $managedImage -ne $ubuntuManagedImage -and $managedImages -ne $windowsManagedImage2022) {
         Write-Host "Found a match, deleting orphaned managed image: $managedImage"
         az image delete --ids $managedImage | Out-Null
     }
